runtime: apply tsconfig customConditions as Node export conditions
TypeScript resolves types out of a package's exports map using
compilerOptions.customConditions, but nothing carried them to the runtime, so a
project on the live-types layout type-checked against a `source` condition and
then loaded the built copy — or failed to resolve at all. nub now reads them and
emits --conditions= NODE_OPTIONS tokens, union'd with nub.jsonc `conditions`.

Measured: --conditions in NODE_OPTIONS reaches BOTH the ESM and CJS resolvers on
18.19.0, 20.19.0 and 26.5.0, so one spawn-time flag covers the whole support
band; no tier split, and it rides through to workers and child node processes.
Skipped in compat mode with every other config-derived flag.

Discovery is anchored at the CWD, matching where nub.jsonc discovery starts and
where Node resolves a bare --import specifier from, and it follows the extends
chain so a condition declared in a shared base config counts.

The get-tsconfig port moves out of nub-native into a new napi-free
crates/nub-tsconfig so the CLI and the addon read one tsconfig rather than two
implementations of the extends chain; the addon keeps only the #[napi] wrapper.
That crate now also owns the pinned JSONC dialect and nesting bound both readers
share, and — unlike the cdylib it came from — it can carry unit tests.
